poll query performance
